fix: G115: integer overflow conversion int64 -> uint16 (gosec)

+++ internal/api/primary_server.go
@@ -8,13 +8,13 @@ import (
 
 type PrimaryServer struct {
 	ID      string `json:"id"`
-	Port    uint16 `json:"port"`
+	Port    int64  `json:"port"`
 	ZoneID  string `json:"zone_id"`
 	Address string `json:"address"`
 }
 
 type CreatePrimaryServerRequest struct {
-	Port    uint16 `json:"port"`
+	Port    int64  `json:"port"`
 	ZoneID  string `json:"zone_id"`
 	Address string `json:"address"`
 }
